backend/projectV1/requirements.txt
aiosignal==1.3.1
asgiref==3.5.2
attrs==23.1.0
autopep8==1.7.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
distro==1.8.0
Django==4.1.3
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
filelock==3.12.2
Flask==2.3.2
frozenlist==1.4.0
grpcio==1.56.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numpy==1.21.2
packaging==23.1
paramiko==3.2.0
Pillow==9.3.0
protobuf==4.23.4
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dotenv==0.21.1
gunicorn==20.1.0
PyYAML==5.4.1
ray==2.6.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
sympy==1.12
tk==0.1.0
toml==0.10.2
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.7.1
urllib3==2.0.3
websocket-client==0.59.0
Werkzeug==2.3.6
psycopg2==2.9.7
gunicorn==20.1.0
scikit-learn==1.3.0
pandas==1.3.3
graphene-django==3.1.5
PyJWT==2.8.0
pyarrow
django-cors-headers==3.10.0
djangorestframework==3.14.0
django-rest-knox==4.2.0
openai
pdfminer.six
pyresparser
pyjwt==2.8.0
channels==4.0.0
channels_redis==4.1.0
redis==5.0.1